[ SW-501 ] fix: Update privacy policy (#4437)
* fix: Update privacy policy

* fix: Remove privacy.md.d.ts

* refactor: Extract CustomLink component

* fix: Update privacy policy style
